refactor: reorganize battery component imports and styles, remove unused styles, and standardize naming conventions for improved clarity and maintainability
