+++ packages/masterfabric-expo-core/src/helpers/battery_helper.ts
