+++ packages/masterfabric-expo-core/src/styles/battery_helper_view.styles.ts
