+++ packages/masterfabric-expo-core/src/styles/battery_status_card.styles.ts
