+++ packages/masterfabric-expo-core/src/styles/device_info_card.styles.ts
