+++ packages/masterfabric-expo-core/src/styles/index.ts
@@ -1,6 +1,7 @@
-export * from './BatteryHelperView.styles';
-export * from './BatteryStatusCard.styles';
-export * from './DeviceInfoCard.styles';
-export * from './LowPowerModeCard.styles';
+export * from './battery_helper_view.styles';
+export * from './battery_status_card.styles';
 export * from './card.styles';
+export * from './device_info_card.styles';
+export * from './low_power_mode_card.styles';
 export * from './screen.styles';
+
