+++ packages/masterfabric-expo-core/src/styles/low_power_mode_card.styles.ts
